Move running/halt state out of Z80CPU class into own class. Move instructions into own class out of Z80Cpu class.  This is enabled by the splitting up of the registers and other classes to allow instructions to operate on registers/I/O/memory etc directly without having to be in the main CPU class. The Z80CPU class now is clean and simple and only handles core instruction execution. The instructions class is still very large but in a much better position to be refactored and split up in the future than previously
