+++ Kmse.Core/Z80/IO/Z80CpuInputOutput.cs
@@ -1,8 +1,8 @@
 ﻿using Kmse.Core.IO;
 using Kmse.Core.Utilities;
+using Kmse.Core.Z80.Model;
 using Kmse.Core.Z80.Registers;
 using Kmse.Core.Z80.Registers.General;
-using Kmse.Core.Z80.Support;
 
 namespace Kmse.Core.Z80.IO;
