+++ Kmse.Core/Z80/Instructions/IZ80CpuInstructions.cs
@@ -1,0 +1,10 @@
+﻿using System.Runtime.CompilerServices;
+using Kmse.Core.Z80.Model;
+
+namespace Kmse.Core.Z80.Instructions
+{
+    public interface IZ80CpuInstructions
+    {
+        Instruction GetInstruction(byte opCode);
+    }
+}
