+++ Kmse.Core/Z80/Memory/Z80CpuMemoryManagement.cs
@@ -1,8 +1,8 @@
 ﻿using Kmse.Core.Memory;
 using Kmse.Core.Utilities;
+using Kmse.Core.Z80.Model;
 using Kmse.Core.Z80.Registers;
 using Kmse.Core.Z80.Registers.General;
-using Kmse.Core.Z80.Support;
 
 namespace Kmse.Core.Z80.Memory;
