+++ Kmse.Core/Z80/Model/CbInstructionModes.cs
@@ -1,0 +1,8 @@
+﻿namespace Kmse.Core.Z80.Model;
+
+public enum CbInstructionModes : byte
+{
+    Normal = 0x00,
+    DD = 0xDD,
+    FD = 0xFD
+}
