+++ Kmse.Core/Z80/Registers/General/IZ80AfRegister.cs
@@ -1,4 +1,4 @@
-﻿using Kmse.Core.Z80.Support;
+﻿using Kmse.Core.Z80.Model;
 
 namespace Kmse.Core.Z80.Registers.General;
