+++ Kmse.Core/Z80/Registers/General/Z80Accumulator.cs
@@ -1,7 +1,7 @@
 ﻿using Kmse.Core.Memory;
 using Kmse.Core.Utilities;
+using Kmse.Core.Z80.Model;
 using Kmse.Core.Z80.Registers.SpecialPurpose;
-using Kmse.Core.Z80.Support;
 
 namespace Kmse.Core.Z80.Registers.General;
