+++ Kmse.Core/Z80/Registers/General/Z80AfRegister.cs
@@ -1,6 +1,6 @@
 ﻿using Kmse.Core.Memory;
 using Kmse.Core.Utilities;
-using Kmse.Core.Z80.Support;
+using Kmse.Core.Z80.Model;
 
 namespace Kmse.Core.Z80.Registers.General;
