+++ Kmse.Core/Z80/Registers/Z8016BitGeneralPurposeRegisterBase.cs
@@ -1,7 +1,7 @@
 ﻿using Kmse.Core.Memory;
 using Kmse.Core.Utilities;
+using Kmse.Core.Z80.Model;
 using Kmse.Core.Z80.Registers.General;
-using Kmse.Core.Z80.Support;
 
 namespace Kmse.Core.Z80.Registers;
