+++ Kmse.Core/Z80/Registers/Z80RegisterBase.cs
@@ -1,6 +1,6 @@
 ﻿using Kmse.Core.Utilities;
+using Kmse.Core.Z80.Model;
 using Kmse.Core.Z80.Registers.General;
-using Kmse.Core.Z80.Support;
 
 namespace Kmse.Core.Z80.Registers;
