+++ Kmse.Core/Z80/Running/IZ80CpuRunningStateManager.cs
@@ -1,0 +1,10 @@
+﻿namespace Kmse.Core.Z80.Running;
+
+public interface IZ80CpuRunningStateManager
+{
+    bool Halted { get; }
+    void Reset();
+    void Halt();
+
+    void ResumeIfHalted();
+}
