+++ Kmse.Core/Z80/Z80CpuManagement.cs
@@ -2,6 +2,7 @@
 using Kmse.Core.Z80.Interrupts;
 using Kmse.Core.Z80.IO;
 using Kmse.Core.Z80.Memory;
+using Kmse.Core.Z80.Running;
 
 namespace Kmse.Core.Z80;
 
@@ -11,4 +12,5 @@ public class Z80CpuManagement
     public IZ80CpuMemoryManagement MemoryManagement { get; set; }
     public IZ80InterruptManagement InterruptManagement { get; set; }
     public IZ80CpuCycleCounter CycleCounter { get; set; }
+    public IZ80CpuRunningStateManager RunningStateManager { get; set; }
 }
